Move max and min chars to the pets constant file
