+++ src/data/constants.ts
@@ -3,7 +3,3 @@ export const defaultDate: Date = new Date(0);
 export const todayDate: Date = new Date();
 
 export const allowedImageTypes = ["image/jpeg", "image/png", "image/webp"];
-
-// can be reused for vet notes and pet diary
-export const notesMinCharacters = 2;
-export const notesMaxCharacters = 1500;
