+++ src/util/validation/validate-diary.test.ts
@@ -1,5 +1,5 @@
 import { validateDiaryContentBody } from "./validate-diary";
-import { notesMaxCharacters, notesMinCharacters } from "~data/constants";
+import { notesMaxCharacters, notesMinCharacters } from "~data/pets/constants";
 
 // Tests generated by Claude Sonnet 4.5
